Fix macOS build: platform detection, include path, ObjC guards
- build_config.h: fix unconditional OS_WIN=1, detect platform correctly
- atomicops.h: remove basic_types.h from non-Windows (include path conflict)
- basic_types.h: move fundamental types before includes
- CMakeLists.txt: remove -I base (shadows system time.h)
- color.h: guard ObjC includes with #ifdef __OBJC__
- win_types.h: add UINT typedef for non-Windows
- thread_checker: add pthread-based non-Windows implementation
- gdiplus_initializer: guard Windows-only code with #ifdef _WIN32
- message_framework/task.h: replace DWORD with uintptr_t
- platform_bitmap_mac.h, platform_font_mac.h: ObjC header fixes

+++ message_framework/task.h
@@ -258,7 +258,7 @@ struct RunnableMethodTraits
 
 private:
 #ifndef NDEBUG
-    DWORD origin_thread_id_;
+    uintptr_t origin_thread_id_;
 #endif
 };
